+++ Source/Readers/NDRMReader/NDRMReader.h
@@ -34,7 +34,7 @@ class NDRMReader : public DataReaderBase
     size_t m_vocabSize;
     size_t m_idfVocabSize;
     size_t m_vectorSize;
-    size_t m_docLengthBinSize;
+    int m_docLengthBinSize;
     size_t m_miniBatchSize;
     size_t m_bytesPerSample;
     size_t m_bytesPerVector;
